requirements.txt
tensorflow
streamlit>=1.32.0
Pillow>=10.0.0
numpy>=1.26.0
matplotlib>=3.8.0
plotly>=5.18.0
pandas>=2.1.0
scikit-learn>=1.4.0
pytest>=8.0.0
